requirements.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing_extensions==4.15.0
